Update index.html
it looks cleaner now. maybe I've actually gotten better at making websites but i doubt it
the side panel is neater, and the page fits the screen for mobile and web.
except for when u open inspect. I'm not changing that DONT INSPECT MY CODE NARC
that's about it, love you